firestore-stdio-python/requirements.txt
mcp>=1.24.0
pydantic==2.12.4
python-json-logger==4.0.0
google-cloud-firestore==2.19.0
python-dotenv>=1.1.0
fastmcp>=2.14.1
